backend/requirements.txt
fastapi
uvicorn
pymongo
dnspython
requests
beautifulsoup4
langchain-core
langchain-google-genai
python-dotenv 
pydantic